feat: studio, item 컨트롤러 구조 작성
