+++ src/main/java/com/example/toucheese_be/domain/item/repository/ItemRepository.java
@@ -1,4 +1,7 @@
-//package com.example.toucheese_be.domain.item.repository;
-//
-//public interface ItemRepository {
-//}
+package com.example.toucheese_be.domain.item.repository;
+
+import com.example.toucheese_be.domain.item.entity.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
