--- src/main/java/com/example/toucheese_be/domain/item/repository/ItemService.java
@@ -1,4 +1,0 @@
-//package com.example.toucheese_be.domain.item.repository;
-//
-//public class ItemService {
-//}
