+++ src/main/java/com/example/toucheese_be/domain/item/service/ItemService.java
@@ -1,0 +1,11 @@
+package com.example.toucheese_be.domain.item.service;
+
+import com.example.toucheese_be.domain.item.repository.ItemRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ItemService {
+    private final ItemRepository itemRepository;
+}
